Remove types and utils from send modal folder
