+++ src/modals/SendModals/SendModalDeployContract.tsx
@@ -28,8 +28,8 @@ import { Client } from '../../contexts/global'
 import { useSendModalContext } from '../../contexts/sendModal'
 import useDappTxData from '../../hooks/useDappTxData'
 import useStateObject from '../../hooks/useStateObject'
-import { DeployContractTxData, TxPreparation } from '../../types/transactions'
-import { hasNoGasErrors, isAmountWithinRange } from '../../utils/transactions'
+import { CheckTxProps, DeployContractTxData, PartialTxData, TxPreparation } from '../../types/transactions'
+import { expectedAmount, hasNoGasErrors, isAmountWithinRange } from '../../utils/transactions'
 import AddressSelectFrom from './AddressSelectFrom'
 import AlphAmountInfoBox from './AlphAmountInfoBox'
 import BuildTxFooterButtons from './BuildTxFooterButtons'
@@ -37,8 +37,6 @@ import BytecodeInput from './BytecodeInput'
 import GasSettingsExpandableSection from './GasSettingsExpandableSection'
 import { ModalInputFields } from './ModalInputFields'
 import SendModal, { TxContext } from './SendModal'
-import { CheckTxProps, PartialTxData } from './types'
-import { expectedAmount } from './utils'
 
 interface DeployContractBuildTxModalContentProps {
   data: PartialTxData<DeployContractTxData, 'fromAddress'>
