+++ src/modals/SendModals/SendModalScript.tsx
@@ -26,8 +26,8 @@ import { Client } from '../../contexts/global'
 import { useSendModalContext } from '../../contexts/sendModal'
 import useDappTxData from '../../hooks/useDappTxData'
 import useStateObject from '../../hooks/useStateObject'
-import { ScriptTxData, TxPreparation } from '../../types/transactions'
-import { hasNoGasErrors, isAmountWithinRange } from '../../utils/transactions'
+import { CheckTxProps, PartialTxData, ScriptTxData, TxPreparation } from '../../types/transactions'
+import { expectedAmount, hasNoGasErrors, isAmountWithinRange } from '../../utils/transactions'
 import AddressSelectFrom from './AddressSelectFrom'
 import AlphAmountInfoBox from './AlphAmountInfoBox'
 import BuildTxFooterButtons from './BuildTxFooterButtons'
@@ -35,8 +35,6 @@ import BytecodeInput from './BytecodeInput'
 import GasSettingsExpandableSection from './GasSettingsExpandableSection'
 import { ModalInputFields } from './ModalInputFields'
 import SendModal, { TxContext } from './SendModal'
-import { CheckTxProps, PartialTxData } from './types'
-import { expectedAmount } from './utils'
 
 interface ScriptBuildTxModalContentProps {
   data: PartialTxData<ScriptTxData, 'fromAddress'>
