+++ src/modals/SendModals/SendModalTransfer.tsx
@@ -28,17 +28,15 @@ import { Client } from '../../contexts/global'
 import { useSendModalContext } from '../../contexts/sendModal'
 import useDappTxData from '../../hooks/useDappTxData'
 import useStateObject from '../../hooks/useStateObject'
-import { TransferTxData, TxPreparation } from '../../types/transactions'
+import { CheckTxProps, PartialTxData, TransferTxData, TxPreparation } from '../../types/transactions'
 import { isAddressValid } from '../../utils/addresses'
-import { hasNoGasErrors, isAmountWithinRange } from '../../utils/transactions'
+import { expectedAmount, hasNoGasErrors, isAmountWithinRange } from '../../utils/transactions'
 import AddressSelectFrom from './AddressSelectFrom'
 import AlphAmountInfoBox from './AlphAmountInfoBox'
 import BuildTxFooterButtons from './BuildTxFooterButtons'
 import GasSettingsExpandableSection from './GasSettingsExpandableSection'
 import { ModalInputFields } from './ModalInputFields'
 import SendModal, { TxContext } from './SendModal'
-import { CheckTxProps, PartialTxData } from './types'
-import { expectedAmount } from './utils'
 
 interface TransferBuildTxModalContentProps {
   data: PartialTxData<TransferTxData, 'fromAddress'>
